frontend/requirements.txt
altair==5.4.1
annotated-types==0.7.0
anthropic==0.32.0
anyio==4.6.2.post1
attrs==21.2.0
Automat==20.2.0
awscli==1.22.34
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
boto3==1.34.151
botocore==1.34.162
cachetools==5.5.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.0
click==8.0.3
cloud-init==24.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
filelock==3.16.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gyp==0.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.26.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.6.1
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
more-itertools==8.10.0
multi_agent_orchestrator==0.0.18
narwhals==1.9.4
netifaces==0.11.0
numpy==2.1.2
oauthlib==3.2.0
olefile==0.46
packaging==24.1
pandas==2.2.3
pexpect==4.8.0
pillow==11.0.0
protobuf==5.28.2
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.32.3
rich==13.9.2
roman==3.3
rsa==4.8
s3transfer==0.10.3
SecretStorage==3.3.1
service-identity==18.1.0
setuptools==59.6.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sos==4.5.6
ssh-import-id==5.11
st-chat-message==0.3.8
streamlit==1.39.0
streamlit-chat==0.1.1
systemd-python==234
tenacity==9.0.0
tokenizers==0.20.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
watchdog==5.0.3
wheel==0.37.1
zipp==1.0.0
zope.interface==5.4.0
